+++ test/fixtures/cookbooks/test/recipes/legacy.rb
@@ -1,3 +1,0 @@
-nodenv_node ['6.11.2'] do
-  user 'notroot'
-end