+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- System install, reusing user install
+### Updated
+- Update readme
+
 ## [0.1.5] - 2018-02-2
 ### Added
 - Nodenv Install resource